refactor(NumberInput): Use partials pattern for styled-components
