+++ packages/react-component-library/src/components/NumberInput/ClearButton.tsx
@@ -1,9 +1,7 @@
 import React from 'react'
 import { IconCancel } from '@royalnavy/icon-library'
-import { selectors } from '@royalnavy/design-tokens'
-import styled, { css } from 'styled-components'
 
-const { color } = selectors
+import { StyledClearButton } from './partials/StyledClearButton'
 
 interface ClearButtonProps {
   isCondensed: boolean
@@ -11,42 +9,13 @@ interface ClearButtonProps {
   onClick: (event: React.MouseEvent<HTMLButtonElement>) => void
 }
 
-interface StyledButtonProps {
-  $isCondensed: boolean
-}
-
-const StyledButton = styled.button<StyledButtonProps>`
-  background-color: ${color('neutral', 'white')};
-  border: none;
-  color: ${color('neutral', '300')};
-  cursor: pointer;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  height: 1rem;
-  right: 54px;
-  padding: unset;
-  display: flex;
-  align-items: center;
-
-  ${({ $isCondensed }) =>
-    $isCondensed &&
-    css`
-      right: 44px;
-
-      & > svg {
-        transform: scale(0.8);
-      }
-    `}
-`
-
 export const ClearButton: React.FC<ClearButtonProps> = ({
   isCondensed,
   isDisabled,
   onClick,
 }) => {
   return (
-    <StyledButton
+    <StyledClearButton
       $isCondensed={isCondensed}
       aria-label="Clear the input value"
       data-testid="number-input-clear"
@@ -55,7 +24,7 @@ export const ClearButton: React.FC<ClearButtonProps> = ({
       type="button"
     >
       <IconCancel />
-    </StyledButton>
+    </StyledClearButton>
   )
 }
