+++ packages/react-component-library/src/components/NumberInput/EndAdornment.tsx
@@ -1,11 +1,8 @@
 import React from 'react'
-import { selectors } from '@royalnavy/design-tokens'
-import styled from 'styled-components'
 
 import { EndAdornmentButton } from './EndAdornmentButton'
 import { END_ADORNMENT_TYPE } from './constants'
-
-const { color } = selectors
+import { StyledNumberInputControls } from './partials/StyledNumberInputControls'
 
 interface EndAdornmentProps {
   isCondensed: boolean
@@ -20,14 +17,6 @@ interface EndAdornmentProps {
   step?: number
   value: number
 }
-
-const StyledNumberInputControls = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  text-align: center;
-  border-left: 1px solid ${color('neutral', '100')};
-`
 
 export const EndAdornment: React.FC<EndAdornmentProps> = ({
   isCondensed,
