+++ packages/react-component-library/src/components/NumberInput/partials/StyledIncreaseButton.tsx
@@ -1,0 +1,24 @@
+import styled, { css } from 'styled-components'
+
+import { StyledButton, StyledButtonProps } from './StyledButton'
+import { StyledDecreaseButton } from './StyledDecreaseButton'
+
+export const StyledIncreaseButton = styled(StyledButton)`
+  & > svg {
+    transform: rotate(180deg);
+  }
+
+  ${({ $isCondensed }) =>
+    $isCondensed &&
+    css`
+      width: 36px;
+
+      & > svg {
+        transform: rotate(180deg) scale(0.7);
+      }
+    `}
+
+  &:focus + ${StyledDecreaseButton} {
+    border-color: transparent;
+  }
+`
