+++ packages/react-component-library/src/components/NumberInput/partials/StyledInputWrapper.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const StyledInputWrapper = styled.div`
+  position: relative;
+  flex-grow: 1;
+`
