+++ packages/react-component-library/src/components/NumberInput/partials/StyledNumberInputUnit.tsx
@@ -1,0 +1,19 @@
+import { selectors } from '@royalnavy/design-tokens'
+import styled from 'styled-components'
+
+const { fontSize } = selectors
+
+interface StyledNumberInputUnitProps {
+  $left: string
+  $top: string
+}
+
+export const StyledNumberInputUnit = styled.span<StyledNumberInputUnitProps>`
+  align-items: center;
+  display: flex;
+  height: 100%;
+  position: absolute;
+  font-size: ${fontSize('base')};
+  left: ${({ $left }) => $left};
+  top: ${({ $top }) => $top};
+`
